fixed std::allocator deprecation warning
Hi,
this is a fix for the very annoying std::allocator-deprecation warnings.
I've changes the evaluation of the fallback-type from greedy to lazy, by quoting it, and only evaluating when needed.
This warning shows up on msvc 14.2 (VS 16.9.1). I don't know if it also occurs with other compilers/std-libs.
I hope it can make it as bugfix still into 1.76

Tobias